setupp.py add translation system
Note that we use egg_info to compile_catalog:

When 'running egg_info', the *.egg-info/SOURCES.txt is generated.
In the SOURCES.txt each file is listed to install after compilation.
If we compile the translation after running egg_info, then the
SOURCES.txt will not have the .mo file. And when we will make a 'python
setup.py install', the .mo will be not included

In the debian packaging process this does not cause any problem
because a 'python setup.py build' is ran before.